External stuff refactor
