+++ src/Gameboard.Api/Features/Game/External/ExternalGamesModels.cs
@@ -1,7 +1,66 @@
 using System;
 using System.Collections.Generic;
+using Gameboard.Api.Data;
 
 namespace Gameboard.Api.Features.Games.External;
+
+
+public sealed class ExternalGameState
+{
+    public required SimpleEntity Game { get; set; }
+    public required ExternalGameStateDeployStatus OverallDeployStatus { get; set; }
+    public required IEnumerable<SimpleEntity> Specs { get; set; }
+    public required DateTimeOffset? StartTime { get; set; }
+    public required DateTimeOffset? EndTime { get; set; }
+    public required bool HasNonStandardSessionWindow { get; set; }
+    public required IEnumerable<ExternalGameStateTeam> Teams { get; set; }
+}
+
+public sealed class ExternalGameStateTeam
+{
+    public required string Id { get; set; }
+    public required string Name { get; set; }
+    public required ExternalGameTeamDeployStatus DeployStatus { get; set; }
+    public required bool IsReady { get; set; }
+    public required IEnumerable<SimpleSponsor> Sponsors { get; set; }
+    public required IEnumerable<ExternalGameStateChallenge> Challenges { get; set; }
+    public required IEnumerable<ExternalGameStatePlayer> Players { get; set; }
+}
+
+public sealed class ExternalGameStateChallenge
+{
+    public required string Id { get; set; }
+    public required bool ChallengeCreated { get; set; }
+    public required bool GamespaceDeployed { get; set; }
+    public required string SpecId { get; set; }
+    public required DateTimeOffset? StartTime { get; set; }
+    public required DateTimeOffset? EndTime { get; set; }
+}
+
+public sealed class ExternalGameStatePlayer
+{
+    public string Id { get; set; }
+    public required string Name { get; set; }
+    public required bool IsCaptain { get; set; }
+    public required SimpleSponsor Sponsor { get; set; }
+    public required ExternalGameStatePlayerStatus Status { get; set; }
+    public required SimpleEntity User { get; set; }
+}
+
+public enum ExternalGameStatePlayerStatus
+{
+    NotConnected,
+    NotReady,
+    Ready
+}
+
+public enum ExternalGameStateDeployStatus
+{
+    NotStarted,
+    PartiallyDeployed,
+    Deploying,
+    Deployed
+}
 
 /// <summary>
 /// This is basically a "frozen" API that pushes data to an external
